requirements.txt
Django==2.2.6
django-mongodb-engine==0.6.0
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
djangotoolbox==1.8.0
dnspython==1.16.0
PyJWT==1.7.1
pymongo==3.9.0
pytz==2019.3
sqlparse==0.3.0
